python/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
networkx>=3.2.0
seaborn>=0.13.0
numpy>=1.26.0
